requirements.txt
fastapi>=0.116,<0.117
uvicorn[standard]>=0.35,<0.36
jinja2>=3.1,<4
sqlalchemy>=2.0,<3
pydantic-settings>=2.10,<3
python-multipart>=0.0.20,<0.0.21
psycopg[binary]>=3.2,<4
